[Decomp][Player] Cohesion split: extract PlayerTaxi helper class to PlayerTaxi.cpp
