+++ src/game/Object/PlayerTaxi.cpp
@@ -1,0 +1,211 @@
+/**
+ * MaNGOS is a full featured server for World of Warcraft, supporting
+ * the following clients: 1.12.x, 2.4.3, 3.3.5a, 4.3.4a and 5.4.8
+ *
+ * Copyright (C) 2005-2025 MaNGOS <https://www.getmangos.eu>
+ *
+ * This program is free software; you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation; either version 2 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program; if not, write to the Free Software
+ * Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
+ *
+ * World of Warcraft, and all World of Warcraft or Warcraft art, images,
+ * and lore are copyrighted by Blizzard Entertainment, Inc.
+ */
+
+
+
+#include "Player.h"
+#include "Language.h"
+#include "Database/DatabaseEnv.h"
+#include "Log.h"
+#include "Opcodes.h"
+#include "World.h"
+#include "WorldPacket.h"
+#include "WorldSession.h"
+#include "UpdateMask.h"
+#include "ObjectMgr.h"
+#include "ObjectAccessor.h"
+#include "DBCStores.h"
+#include "MapManager.h"
+
+/**
+ * @brief Initializes known taxi nodes for a newly created player.
+ *
+ * @param race The player race id.
+ * @param level Unused player level.
+ */
+void PlayerTaxi::InitTaxiNodesForLevel(uint32 race, uint32 level)
+{
+    // race specific initial known nodes: capital and taxi hub masks
+    switch (race)
+    {
+        case RACE_HUMAN:    SetTaximaskNode(2);  break;     // Human
+        case RACE_ORC:      SetTaximaskNode(23); break;     // Orc
+        case RACE_DWARF:    SetTaximaskNode(6);  break;     // Dwarf
+        case RACE_NIGHTELF: SetTaximaskNode(26);
+            SetTaximaskNode(27); break;     // Night Elf
+        case RACE_UNDEAD:   SetTaximaskNode(11); break;     // Undead
+        case RACE_TAUREN:   SetTaximaskNode(22); break;     // Tauren
+        case RACE_GNOME:    SetTaximaskNode(6);  break;     // Gnome
+        case RACE_TROLL:    SetTaximaskNode(23); break;     // Troll
+        case RACE_BLOODELF: SetTaximaskNode(82); break;     // Blood Elf
+        case RACE_DRAENEI:  SetTaximaskNode(94); break;     // Draenei
+    }
+
+    // new continent starting masks (It will be accessible only at new map)
+    switch (Player::TeamForRace(race))
+    {
+        case ALLIANCE: SetTaximaskNode(100); break;
+        case HORDE:    SetTaximaskNode(99);  break;
+        default: break;
+    }
+    // level dependent taxi hubs
+    if (level >= 68)
+    {
+        SetTaximaskNode(213);                               // Shattered Sun Staging Area
+    }
+}
+
+/**
+ * @brief Loads the known taxi-node bitmask from a serialized string.
+ *
+ * @param data The serialized taxi mask data.
+ */
+void PlayerTaxi::LoadTaxiMask(const char* data)
+{
+    Tokens tokens = StrSplit(data, " ");
+
+    int index;
+    Tokens::iterator iter;
+    for (iter = tokens.begin(), index = 0; (index < TaxiMaskSize) && (iter != tokens.end()); ++iter, ++index)
+    {
+        // load and set bits only for existing taxi nodes
+        m_taximask[index] = sTaxiNodesMask[index] & uint32(atol((*iter).c_str()));
+    }
+}
+
+/**
+ * @brief Appends the player's taxi-node mask to a packet buffer.
+ *
+ * @param data The destination packet buffer.
+ * @param all true to append all existing nodes instead of only known nodes.
+ */
+void PlayerTaxi::AppendTaximaskTo(ByteBuffer& data, bool all)
+{
+    if (all)
+    {
+        for (uint8 i = 0; i < TaxiMaskSize; ++i)
+        {
+            data << uint32(sTaxiNodesMask[i]);               // all existing nodes
+        }
+    }
+    else
+    {
+        for (uint8 i = 0; i < TaxiMaskSize; ++i)
+        {
+            data << uint32(m_taximask[i]);                   // known nodes
+        }
+    }
+}
+
+/**
+ * @brief Loads active taxi destinations from a serialized path string.
+ *
+ * @param values The serialized taxi destination list.
+ * @param team The player's faction team.
+ * @return true if the taxi route is valid; otherwise, false.
+ */
+bool PlayerTaxi::LoadTaxiDestinationsFromString(const std::string& values, Team team)
+{
+    ClearTaxiDestinations();
+
+    Tokens tokens = StrSplit(values, " ");
+
+    for (Tokens::iterator iter = tokens.begin(); iter != tokens.end(); ++iter)
+    {
+        uint32 node = uint32(atol(iter->c_str()));
+        AddTaxiDestination(node);
+    }
+
+    if (m_TaxiDestinations.empty())
+    {
+        return true;
+    }
+
+    // Check integrity
+    if (m_TaxiDestinations.size() < 2)
+    {
+        return false;
+    }
+
+    for (size_t i = 1; i < m_TaxiDestinations.size(); ++i)
+    {
+        uint32 cost;
+        uint32 path;
+        sObjectMgr.GetTaxiPath(m_TaxiDestinations[i - 1], m_TaxiDestinations[i], path, cost);
+        if (!path)
+        {
+            return false;
+        }
+    }
+
+    // can't load taxi path without mount set (quest taxi path?)
+    if (!sObjectMgr.GetTaxiMountDisplayId(GetTaxiSource(), team, true))
+    {
+        return false;
+    }
+
+    return true;
+}
+
+/**
+ * @brief Serializes the current taxi destination list.
+ *
+ * @return The serialized taxi destination string.
+ */
+std::string PlayerTaxi::SaveTaxiDestinationsToString()
+{
+    if (m_TaxiDestinations.empty())
+    {
+        return "";
+    }
+
+    std::ostringstream ss;
+
+    for (size_t i = 0; i < m_TaxiDestinations.size(); ++i)
+    {
+        ss << m_TaxiDestinations[i] << " ";
+    }
+
+    return ss.str();
+}
+
+/**
+ * @brief Gets the taxi path id for the current first route segment.
+ *
+ * @return The current taxi path id, or 0 if no valid route exists.
+ */
+uint32 PlayerTaxi::GetCurrentTaxiPath() const
+{
+    if (m_TaxiDestinations.size() < 2)
+    {
+        return 0;
+    }
+
+    uint32 path;
+    uint32 cost;
+
+    sObjectMgr.GetTaxiPath(m_TaxiDestinations[0], m_TaxiDestinations[1], path, cost);
+
+    return path;
+}
